Remove all type annotations from skeletons
With zsh extended globbing:

    sed -i 's/::\w*//g' **/(^example).jl~**/runtests.jl

+++ exercises/roman-numerals/roman-numerals.jl
@@ -1,3 +1,3 @@
-function to_roman(number::Integer)
+function to_roman(number)
 
 end
